fix make vet
